Fixed Add Contract and Materials Needed

+++ UI/dashbord.php
@@ -50,25 +50,30 @@
             <tbody>
              <?php
 
-               $userlist = $user->getAllcontracts();
+               $contractList = $user->getAllcontracts();
 
-               foreach($userlist as $user){
-                  $contract_id = $user['contract_id'];
-            ?>
+               if($contractList){
+                  foreach($contractList as $contract){
+                     $contract_id = $contract['contract_id'];
+                  ?>
 
-                <tr>
-                <th scope='row'><?php echo $user['contract_id'] ?></th>
-                <td><?php echo $user['client_name'] ?></td>
-                <td><?php echo $user['start_date'] ?></td>
-                <td><?php echo $user['deadline_date'] ?></td>
-                <td><?php echo $user['claim_day'] ?></td>
-                <td><a href='details.php?contract_id=<?php echo $user['contract_id'] ?>'><button type='button' class='btn btn-outline-info'><i class='fas fa-edit'>Detail</i></button></a></td>
-                <td><a href='editContract.php'><button type='button' class='btn btn-outline-warning'>Edit</button></a></td>
-                <td><a href='../deleteAction.php?actiontype=delete&contract_id=$contract_id'><button type='button' name='delete' class='btn btn-outline-danger'>Delete</button></a></td>
-                </tr>
-                
-                 
-               <?php } ?>
+                  <tr>
+                  <th scope='row'><?php echo $contract['contract_id'] ?></th>
+                  <td><?php echo $contract['client_name'] ?></td>
+                  <td><?php echo $contract['start_date'] ?></td>
+                  <td><?php echo $contract['deadline_date'] ?></td>
+                  <td><?php echo $contract['claim_day'] ?></td>
+                  <td><a href='details.php?contract_id=<?php echo $contract['contract_id'] ?>'><button type='button' class='btn btn-outline-info'><i class='fas fa-edit'>Detail</i></button></a></td>
+                  <td><a href='editContract.php'><button type='button' class='btn btn-outline-warning'>Edit</button></a></td>
+                  <td><a href='../deleteAction.php?actiontype=delete&contract_id=$contract_id'><button type='button' name='delete' class='btn btn-outline-danger'>Delete</button></a></td>
+                  </tr>
+                  
+                  
+                  <?php }
+               }else{
+                  echo "<td colspan=8 class='text-center text-danger font-weight-bold'>No Records Found</td>";
+               }
+                ?>
             </tbody>
                   
             </table>
